Explicitly call de-duplication helper for Zoom connectors.
This is required due to the granularity of filtering provided by the Zoom API being at odds with the granularity of the events returned. In order to prevent duplicates, a new utility has been added to allow explicit handling of this edge case.

This commit also aligns applicable connectors to a 7-day look back.