Add configurable delay for Google and Github. (#20)
* Add configurable collection delay for Github.

In-line with feedback in #17 Github also makes some logs available before others. As a result, a delay configuration has been added to allow operations to preference consistency over collection time. In future, more in-depth retrospection and deduplication will be added which should remove the need for this parameter.

However, for now, this enables the operator to modify delay based on their appeteite for potentially missed logs - due to no lag timeframes being provided by Github, and log entries being back-filled based on this unknown interval.

* Add configurable collection delay for GSuite.

In-line with feedback in #17, Google activities have a published 'lag' time before which logs cannot be considered consistent. As a result, a delay configuration has been added to allow operations to preference consistency over collection time. In future, more in-depth retrospection and deduplication will be added which should remove the need for this parameter.

* Add delay to GSuite activities configuration example.

* Github query does not require 'AND'

* Keep mypy happy.

Non-optional fields are explicitly marked as 'type: ignore' as although they're optional due to hydration from environment variables without defaults, they're enforced by validation and later tests that they're not None.