Move to pyproject.toml. Add processors, documentation updates. (#26)
* Add processors. Move to pyproject.toml.

* Cache calls to quote_aware_split.

* Update processor base class to allow finalize only

This was possible prior to this commit, but required a stub method for process to be implemented.

* Add local file secret backend.

* Remove caching to prevent unexpected mutation.

* Documentation updates.

* Version bump for new release.

* Update build process for pyproject.toml

* Only write processed data if processors are configured.

By default both a processed and raw output stream are configured to simplify configuration. However, this means that if no processors are used logs will be written twice. This pull-request prevents that, and instead only write out data if processors are configured.

* Fix up logger error.

* Update grove/outputs/__init__.py

Co-authored-by: Jonas Plum <cugu@users.noreply.github.com>

* Address pull-request comments.

---------

Co-authored-by: Jonas Plum <cugu@users.noreply.github.com>