[1.4.0] Add remote HTTP output handler. (#48)
* Add remote HTTP output handler.

This commit adds the initial remote HTTP output handler. This allows logs to be submitted to a configured HTTP / HTTPS endpoint via HTTP POST in ND-JSON format.

* Fix test to match fixture after change.

* Fix bug in update_path.

This bug results in a path that doesn't exist raising a KeyError. This should not happen, as if the key doesn't exist, we don't need to do anything.

* Apply suggestions from code review

Per Conway - spelling issue.