Add Snowflake login, query, and session history connector. (#59)
* Allow passing Python native data-types to serialize.

This allows passing Datetime objects, and other types which are not serializable to JSON directly. This may not work as expected to all data-types, but will suffice until such a time where a custom JSON Encoder is required to handle something off-the-wall.

* Add Snowflake login, query, and session history connector.

* Additional error handling for Snowflake.

* Remove PKCS#8 keys from test fixtures.

Although these were unique and not used for anything, these have been removed to avoid the opportunity for someone to accidentally use these as real keys in the future. Although the chance of this happening is very low, it's best to eliminate this vector entirely.